+++ clients/drs.d.ts
@@ -60,6 +60,14 @@ declare class Drs extends Service {
    */
   deleteJob(callback?: (err: AWSError, data: Drs.Types.DeleteJobResponse) => void): Request<Drs.Types.DeleteJobResponse, AWSError>;
   /**
+   * Deletes a resource launch action.
+   */
+  deleteLaunchAction(params: Drs.Types.DeleteLaunchActionRequest, callback?: (err: AWSError, data: Drs.Types.DeleteLaunchActionResponse) => void): Request<Drs.Types.DeleteLaunchActionResponse, AWSError>;
+  /**
+   * Deletes a resource launch action.
+   */
+  deleteLaunchAction(callback?: (err: AWSError, data: Drs.Types.DeleteLaunchActionResponse) => void): Request<Drs.Types.DeleteLaunchActionResponse, AWSError>;
+  /**
    * Deletes a single Launch Configuration Template by ID.
    */
   deleteLaunchConfigurationTemplate(params: Drs.Types.DeleteLaunchConfigurationTemplateRequest, callback?: (err: AWSError, data: Drs.Types.DeleteLaunchConfigurationTemplateResponse) => void): Request<Drs.Types.DeleteLaunchConfigurationTemplateResponse, AWSError>;
@@ -228,6 +236,14 @@ declare class Drs extends Service {
    */
   listExtensibleSourceServers(callback?: (err: AWSError, data: Drs.Types.ListExtensibleSourceServersResponse) => void): Request<Drs.Types.ListExtensibleSourceServersResponse, AWSError>;
   /**
+   * Lists resource launch actions.
+   */
+  listLaunchActions(params: Drs.Types.ListLaunchActionsRequest, callback?: (err: AWSError, data: Drs.Types.ListLaunchActionsResponse) => void): Request<Drs.Types.ListLaunchActionsResponse, AWSError>;
+  /**
+   * Lists resource launch actions.
+   */
+  listLaunchActions(callback?: (err: AWSError, data: Drs.Types.ListLaunchActionsResponse) => void): Request<Drs.Types.ListLaunchActionsResponse, AWSError>;
+  /**
    * Returns an array of staging accounts for existing extended source servers.
    */
   listStagingAccounts(params: Drs.Types.ListStagingAccountsRequest, callback?: (err: AWSError, data: Drs.Types.ListStagingAccountsResponse) => void): Request<Drs.Types.ListStagingAccountsResponse, AWSError>;
@@ -243,6 +259,14 @@ declare class Drs extends Service {
    * List all tags for your Elastic Disaster Recovery resources.
    */
   listTagsForResource(callback?: (err: AWSError, data: Drs.Types.ListTagsForResourceResponse) => void): Request<Drs.Types.ListTagsForResourceResponse, AWSError>;
+  /**
+   * Puts a resource launch action.
+   */
+  putLaunchAction(params: Drs.Types.PutLaunchActionRequest, callback?: (err: AWSError, data: Drs.Types.PutLaunchActionResponse) => void): Request<Drs.Types.PutLaunchActionResponse, AWSError>;
+  /**
+   * Puts a resource launch action.
+   */
+  putLaunchAction(callback?: (err: AWSError, data: Drs.Types.PutLaunchActionResponse) => void): Request<Drs.Types.PutLaunchActionResponse, AWSError>;
   /**
    * WARNING: RetryDataReplication is deprecated. Causes the data replication initiation sequence to begin immediately upon next Handshake for the specified Source Server ID, regardless of when the previous initiation started. This command will work only if the Source Server is stalled or is in a DISCONNECTED or STOPPED state. 
    */
@@ -492,6 +516,10 @@ declare namespace Drs {
      */
     licensing?: Licensing;
     /**
+     * Whether we want to activate post-launch actions.
+     */
+    postLaunchEnabled?: Boolean;
+    /**
      * Request to associate tags during creation of a Launch Configuration Template.
      */
     tags?: TagsMap;
@@ -691,6 +719,12 @@ declare namespace Drs {
     jobID: JobID;
   }
   export interface DeleteJobResponse {
+  }
+  export interface DeleteLaunchActionRequest {
+    actionId: LaunchActionId;
+    resourceId: LaunchActionResourceId;
+  }
+  export interface DeleteLaunchActionResponse {
   }
   export interface DeleteLaunchConfigurationTemplateRequest {
     /**
@@ -1047,6 +1081,7 @@ declare namespace Drs {
   export type FailbackLaunchType = "RECOVERY"|"DRILL"|string;
   export type FailbackReplicationError = "AGENT_NOT_SEEN"|"FAILBACK_CLIENT_NOT_SEEN"|"NOT_CONVERGING"|"UNSTABLE_NETWORK"|"FAILED_TO_ESTABLISH_RECOVERY_INSTANCE_COMMUNICATION"|"FAILED_TO_DOWNLOAD_REPLICATION_SOFTWARE_TO_FAILBACK_CLIENT"|"FAILED_TO_CONFIGURE_REPLICATION_SOFTWARE"|"FAILED_TO_PAIR_AGENT_WITH_REPLICATION_SOFTWARE"|"FAILED_TO_ESTABLISH_AGENT_REPLICATOR_SOFTWARE_COMMUNICATION"|"FAILED_GETTING_REPLICATION_STATE"|"SNAPSHOTS_FAILURE"|"FAILED_TO_CREATE_SECURITY_GROUP"|"FAILED_TO_LAUNCH_REPLICATION_SERVER"|"FAILED_TO_BOOT_REPLICATION_SERVER"|"FAILED_TO_AUTHENTICATE_WITH_SERVICE"|"FAILED_TO_DOWNLOAD_REPLICATION_SOFTWARE"|"FAILED_TO_CREATE_STAGING_DISKS"|"FAILED_TO_ATTACH_STAGING_DISKS"|"FAILED_TO_PAIR_REPLICATION_SERVER_WITH_AGENT"|"FAILED_TO_CONNECT_AGENT_TO_REPLICATION_SERVER"|"FAILED_TO_START_DATA_TRANSFER"|string;
   export type FailbackState = "FAILBACK_NOT_STARTED"|"FAILBACK_IN_PROGRESS"|"FAILBACK_READY_FOR_LAUNCH"|"FAILBACK_COMPLETED"|"FAILBACK_ERROR"|"FAILBACK_NOT_READY_FOR_LAUNCH"|"FAILBACK_LAUNCH_STATE_NOT_AVAILABLE"|string;
+  export type FailureReason = string;
   export interface GetFailbackReplicationConfigurationRequest {
     /**
      * The ID of the Recovery Instance whose failback replication configuration should be returned.
@@ -1200,6 +1235,92 @@ declare namespace Drs {
   export type LargeBoundedString = string;
   export type LastLaunchResult = "NOT_STARTED"|"PENDING"|"SUCCEEDED"|"FAILED"|string;
   export type LastLaunchType = "RECOVERY"|"DRILL"|string;
+  export interface LaunchAction {
+    /**
+     * Launch action code.
+     */
+    actionCode?: SsmDocumentName;
+    actionId?: LaunchActionId;
+    actionVersion?: LaunchActionVersion;
+    /**
+     * Whether the launch action is active.
+     */
+    active?: Boolean;
+    category?: LaunchActionCategory;
+    description?: LaunchActionDescription;
+    name?: LaunchActionName;
+    /**
+     * Whether the launch will not be marked as failed if this action fails.
+     */
+    optional?: Boolean;
+    order?: LaunchActionOrder;
+    parameters?: LaunchActionParameters;
+    /**
+     * Launch action type.
+     */
+    type?: LaunchActionType;
+  }
+  export type LaunchActionCategory = "MONITORING"|"VALIDATION"|"CONFIGURATION"|"SECURITY"|"OTHER"|string;
+  export type LaunchActionDescription = string;
+  export type LaunchActionId = string;
+  export type LaunchActionIds = LaunchActionId[];
+  export type LaunchActionName = string;
+  export type LaunchActionOrder = number;
+  export interface LaunchActionParameter {
+    /**
+     * Type.
+     */
+    type?: LaunchActionParameterType;
+    /**
+     * Value.
+     */
+    value?: LaunchActionParameterValue;
+  }
+  export type LaunchActionParameterName = string;
+  export type LaunchActionParameterType = "SSM_STORE"|"DYNAMIC"|string;
+  export type LaunchActionParameterValue = string;
+  export type LaunchActionParameters = {[key: string]: LaunchActionParameter};
+  export type LaunchActionResourceId = string;
+  export interface LaunchActionRun {
+    /**
+     * Action.
+     */
+    action?: LaunchAction;
+    /**
+     * Failure reason.
+     */
+    failureReason?: FailureReason;
+    /**
+     * Run Id.
+     */
+    runId?: LaunchActionRunId;
+    /**
+     * Run status.
+     */
+    status?: LaunchActionRunStatus;
+  }
+  export type LaunchActionRunId = string;
+  export type LaunchActionRunStatus = "IN_PROGRESS"|"SUCCEEDED"|"FAILED"|string;
+  export type LaunchActionRuns = LaunchActionRun[];
+  export type LaunchActionType = "SSM_AUTOMATION"|"SSM_COMMAND"|string;
+  export type LaunchActionVersion = string;
+  export type LaunchActions = LaunchAction[];
+  export interface LaunchActionsRequestFilters {
+    /**
+     * Launch actions Ids.
+     */
+    actionIds?: LaunchActionIds;
+  }
+  export interface LaunchActionsStatus {
+    /**
+     * List of post launch action status.
+     */
+    runs?: LaunchActionRuns;
+    /**
+     * Time where the AWS Systems Manager was detected as running on the launched instance.
+     */
+    ssmAgentDiscoveryDatetime?: ISO8601DatetimeString;
+  }
   export interface LaunchConfiguration {
     /**
      * Whether we should copy the Private IP of the Source Server to the Recovery Instance.
@@ -1225,6 +1346,10 @@ declare namespace Drs {
      * The name of the launch configuration.
      */
     name?: SmallBoundedString;
+    /**
+     * Whether we want to activate post-launch actions for the Source Server.
+     */
+    postLaunchEnabled?: Boolean;
     /**
      * The ID of the Source Server for this launch configuration.
      */
@@ -1263,6 +1388,10 @@ declare namespace Drs {
      * Licensing.
      */
     licensing?: Licensing;
+    /**
+     * Post-launch actions activated.
+     */
+    postLaunchEnabled?: Boolean;
     /**
      * Tags of the Launch Configuration Template.
      */
@@ -1350,6 +1479,31 @@ declare namespace Drs {
     items?: StagingSourceServersList;
     /**
      * The token of the next extensible source server to retrieve.
+     */
+    nextToken?: PaginationToken;
+  }
+  export interface ListLaunchActionsRequest {
+    /**
+     * Filters to apply when listing resource launch actions.
+     */
+    filters?: LaunchActionsRequestFilters;
+    /**
+     * Maximum amount of items to return when listing resource launch actions.
+     */
+    maxResults?: MaxResultsType;
+    /**
+     * Next token to use when listing resource launch actions.
+     */
+    nextToken?: PaginationToken;
+    resourceId: LaunchActionResourceId;
+  }
+  export interface ListLaunchActionsResponse {
+    /**
+     * List of resource launch actions.
+     */
+    items?: LaunchActions;
+    /**
+     * Next token returned when listing resource launch actions.
      */
     nextToken?: PaginationToken;
   }
@@ -1454,6 +1608,10 @@ declare namespace Drs {
   export type ParticipatingResources = ParticipatingResource[];
   export interface ParticipatingServer {
     /**
+     * The post-launch action runs of a participating server.
+     */
+    launchActionsStatus?: LaunchActionsStatus;
+    /**
      * The launch status of a participating server.
      */
     launchStatus?: LaunchStatus;
@@ -1468,6 +1626,54 @@ declare namespace Drs {
   }
   export type ParticipatingServers = ParticipatingServer[];
   export type PositiveInteger = number;
+  export interface PutLaunchActionRequest {
+    /**
+     * Launch action code.
+     */
+    actionCode: SsmDocumentName;
+    actionId: LaunchActionId;
+    actionVersion: LaunchActionVersion;
+    /**
+     * Whether the launch action is active.
+     */
+    active: Boolean;
+    category: LaunchActionCategory;
+    description?: LaunchActionDescription;
+    name: LaunchActionName;
+    /**
+     * Whether the launch will not be marked as failed if this action fails.
+     */
+    optional: Boolean;
+    order: LaunchActionOrder;
+    parameters?: LaunchActionParameters;
+    resourceId: LaunchActionResourceId;
+  }
+  export interface PutLaunchActionResponse {
+    /**
+     * Launch action code.
+     */
+    actionCode?: SsmDocumentName;
+    actionId?: LaunchActionId;
+    actionVersion?: LaunchActionVersion;
+    /**
+     * Whether the launch action is active.
+     */
+    active?: Boolean;
+    category?: LaunchActionCategory;
+    description?: LaunchActionDescription;
+    name?: LaunchActionName;
+    /**
+     * Whether the launch will not be marked as failed if this action fails.
+     */
+    optional?: Boolean;
+    order?: LaunchActionOrder;
+    parameters?: LaunchActionParameters;
+    resourceId?: LaunchActionResourceId;
+    /**
+     * Launch action type.
+     */
+    type?: LaunchActionType;
+  }
   export interface RecoveryInstance {
     /**
      * The ARN of the Recovery Instance.
@@ -2117,6 +2323,7 @@ declare namespace Drs {
   export type SourceServerID = string;
   export type SourceServerIDs = SourceServerID[];
   export type SourceServersList = SourceServer[];
+  export type SsmDocumentName = string;
   export interface StagingArea {
     /**
      * Shows an error message that occurred when DRS tried to access the staging source server. In this case StagingArea$status will have value EXTENSION_ERROR
@@ -2363,6 +2570,10 @@ declare namespace Drs {
      */
     name?: SmallBoundedString;
     /**
+     * Whether we want to enable post-launch actions for the Source Server.
+     */
+    postLaunchEnabled?: Boolean;
+    /**
      * The ID of the Source Server that we want to retrieve a Launch Configuration for.
      */
     sourceServerID: SourceServerID;
@@ -2396,6 +2607,10 @@ declare namespace Drs {
      * Licensing.
      */
     licensing?: Licensing;
+    /**
+     * Whether we want to activate post-launch actions.
+     */
+    postLaunchEnabled?: Boolean;
     /**
      * Target instance type right-sizing method.
      */